==> notebooks/step0_patient_analysis.ipynb <==
# Patient Analysis

This script analysis certain information about the patients.

import numpy as np
import pandas as pd
import scipy as sp
from scipy import io as sio
from scipy import signal as sps
from scipy import linalg as spl
import matplotlib.pyplot as plt
import seaborn as sns
from os.path import join as osj

import pickle
import wfdb


## Get patient infos

ROOT = osj("..", "physionet.org/files/mitdb/1.0.0")
RECORDS = osj(ROOT, "RECORDS")

def get_ecg_info(patient_ids):
    """
    The MIT-BIH data additionally contains ecg info, providing additional information, such as the age, sex, gender, and comments.
    """
    info = {}
    for id_ in patient_ids:
        _, info_ = wfdb.io.rdsamp(osj(ROOT, str(id_)))
        info[id_] = info_["comments"]
    return info

patient_ids = pd.read_csv(osj("..", "files", "patient_ids.csv"), header=None).to_numpy().reshape(-1)
ecg_info = get_ecg_info(patient_ids)

ecg_info

dict_patient_info = {}

for patient_id, entries in ecg_info.items():

    # Get age and gender
    first_entry = [part.strip() for part in entries[0].split(' ')]
    age = int(first_entry[0])
    gender = first_entry[1]

    # Get medication
    medication_raw = entries[1]
    medication = [med.strip() for med in medication_raw.split(',')]

    # Get comments
    comments = entries[2:] if len(entries) > 2 else []
    comment_string = " ".join(comments)

    # Store in new dictionary
    dict_patient_info[patient_id] = {
        "age": age,
        "gender": gender,
        "medication": medication,
        "comment": comment_string
    }

dict_patient_info

def save_patient_infos(dict_patient_info):
    with open(osj("..", "files", "patient_infos.pkl"), "wb") as f:
        pickle.dump(dict_patient_info, f)

# save_patient_infos(dict_patient_info)

## Get patient files

patient_ids = pd.read_csv(osj("..", "files", "patient_ids.csv"), header=None).to_numpy().reshape(-1)

valid_patients = pd.read_csv(osj("..", "files", "valid_patients.csv"), header=None).to_numpy().reshape(-1)
paced_patients = pd.read_csv(osj("..", "files", "paced_patients.csv"), header=None).to_numpy().reshape(-1)
excluded_patients = pd.read_csv(osj("..", "files", "excluded_patients.csv"), header=None).to_numpy().reshape(-1)
class_counts = pd.read_csv(osj("..", "files", "patient_beat_class_counts.csv"), header=None).to_numpy().reshape(-1)

patient_infos = pickle.load(open(osj("..", "files", "patient_infos.pkl"), "rb"))

len(patient_ids), len(paced_patients), len(excluded_patients), len(valid_patients)

valid_patients

## Get class statistics

columns   = class_counts[:3]
data_only = class_counts[3:]

triplets = [data_only[i:i+3] for i in range(0, len(data_only), 3)]
df_class_counts = pd.DataFrame(triplets, columns=columns)

df_class_counts["patient_id"] = df_class_counts["patient_id"].astype(int)
df_class_counts["num_healthy"] = df_class_counts["num_healthy"].astype(int)
df_class_counts["num_arrhythmia"] = df_class_counts["num_arrhythmia"].astype(int)

df_class_counts

valid_class_counts = df_class_counts[df_class_counts["patient_id"].isin(valid_patients)]

# get sum of arrhythmia and healthy beats
total_healthy = valid_class_counts["num_healthy"].sum()
total_arrhythmia = valid_class_counts["num_arrhythmia"].sum()

# get percentage of arrhythmia and healthy beats
total = total_healthy + total_arrhythmia
perc_healthy = total_healthy / total * 100
perc_arrhythmia = total_arrhythmia / total * 100

print(f"Total number of valid beats: {total}")
print(f"Sum of healthy beats: {total_healthy} ({perc_healthy:.2f}%)")
print(f"Sum of arrhythmic beats: {total_arrhythmia} ({perc_arrhythmia:.2f}%)")
